DBRef cleanup
* Removed DBRef and introduced ForeignObjectId
* For some reason DataMapper::Mongo::Property was a class, using a
  module for now
* ForeignObjectId and ObjectId now share a module for common code
* Adapted all specs to pass this changes
