+++ lib/dm-mongo-adapter.rb
@@ -8,7 +8,8 @@ dir = Pathname(__FILE__).dirname.expand_path + 'dm-mongo-adapter'
 require dir + 'query'
 require dir + 'query' + 'java_script'
 
-require dir + 'property' + 'db_ref'
+require dir + 'property' + 'bson_object_id'
+require dir + 'property' + 'foreign_object_id'
 require dir + 'property' + 'object_id'
 require dir + 'property' + 'array'
 require dir + 'property' + 'hash'
