+++ lib/dm-mongo-adapter/property/array.rb
@@ -1,6 +1,6 @@
 module DataMapper
   module Mongo
-    class Property
+    module Property
       class Array < DataMapper::Property::Object
         include DataMapper::Property::PassThroughLoadDump
         primitive ::Array
