+++ lib/dm-mongo-adapter/property/object_id.rb
@@ -1,6 +1,6 @@
 module DataMapper
   module Mongo
-    class Property
+    module Property
       # Each object (document) stored in Mongo DB has an _id field as its
       # first attribute.  This is an object id.  It must be unique for each
       # member of a collection (this is enforced if the collection has an _id
@@ -14,10 +14,17 @@ module DataMapper
       # @see http://www.mongodb.org/display/DOCS/Object+IDs
       #
       # @api public
-      class ObjectId < DBRef
+      class ObjectId < DataMapper::Property::Object
+        include BsonObjectId
+        primitive ::BSON::ObjectId
         key true
         field '_id'
       end # ObjectId
+
+      # @api private
+      def to_child_key
+        ForeignObjectId
+      end
     end # Property
   end # Mongo
 end # DataMapper
