+++ spec/public/properties/object_id_spec.rb
@@ -11,5 +11,9 @@ describe DataMapper::Mongo::Property::ObjectId do
     @property       = User.properties[:id]
   end
 
+  it 'should have ForeignObjectId as #child_key' do
+    @property.child_key.should == DataMapper::Mongo::Property::ForeignObjectId
+  end
+
   it_should_behave_like 'An ObjectId Type'
 end
